+++ sink_tostdout_test.go
@@ -79,3 +79,45 @@ func TestToStdoutWritesToProgramStdout(t *testing.T) {
 
 	assert.Equal(t, expectedResult, actualResult)
 }
+
+func TestToStdoutWritesToTheTraceOutput(t *testing.T) {
+
+	// ----------------------------------------------------------------
+	// setup your test
+
+	// hijack stdout for this test
+	oldStdout := os.Stdout
+	_, w, _ := os.Pipe()
+	os.Stdout = w
+
+	// clean up after ourselves
+	defer func() { os.Stdout = oldStdout }()
+
+	expectedResult := `+ Echo("hello world")
++ => Echo("hello world")
++ p.Stdout> hello world
++ ToStdout()
++ os.Stdout> hello world
+`
+	dest := NewDest()
+	GetShellOptions().EnableTrace(dest)
+
+	// clean up after ourselves
+	defer GetShellOptions().DisableTrace()
+
+	list := NewList(
+		Echo("hello world"),
+		ToStdout(),
+	)
+
+	// ----------------------------------------------------------------
+	// perform the change
+
+	list.Exec()
+	actualResult := dest.String()
+
+	// ----------------------------------------------------------------
+	// test the results
+
+	assert.Equal(t, expectedResult, actualResult)
+}